feat: support autofixing files
This change adds support for autofixing files. This is achieved by
expanding the script that runs as the Docker entrypoint to run the
`kubescape fix` command if a user requests it. It also validates that
the prerequisites to fix the files are met: a scan must include a JSON
output.